Convert old damage system to damage component

+++ src/main/java/net/mcreator/element/converter/v2025_4/ItemToBedrockConverter.java
@@ -47,7 +47,7 @@ public class ItemToBedrockConverter implements IConverter {
 			beitem.useDuration = (double) item.useDuration / 20;
 			beitem.maxDurability = item.damageCount;
 			beitem.enableMeleeDamage = item.enableMeleeDamage;
-			beitem.damageVsEntity = item.damageVsEntity;
+			beitem.damageVsEntity = (int) item.damageVsEntity;
 
 			if (itemDefinition.has("glowCondition")) {
 				JsonObject itemGlowCondition = itemDefinition.getAsJsonObject("glowCondition");
